Move CoreDataFeedStore's `FeedStore` and `FeedImageStore` implementations to separate files
