+++ EssentialFeed/Feed Cache/CoreData/CoreDataFeedStore.swift
@@ -1,17 +1,17 @@
 import Foundation
 import CoreData
 
-public class CoreDataFeedStore: FeedStore {
+public class CoreDataFeedStore {
     private static let modelName: String = "FeedStore"
     private static let model = NSManagedObjectModel.with(name: modelName, in: Bundle(for: CoreDataFeedStore.self))
 
     private let container: NSPersistentContainer
     private let context: NSManagedObjectContext
 
-    enum StoreError: Error {
-            case modelNotFound
-            case failedToLoadPersistentContainer(Error)
-        }
+    public enum StoreError: Error {
+        case modelNotFound
+        case failedToLoadPersistentContainer(Error)
+    }
 
     public init(storeURL: URL) throws {
         guard let model = CoreDataFeedStore.model else {
@@ -28,84 +28,5 @@ public class CoreDataFeedStore: FeedStore {
 
     func perform(_ block: @escaping (NSManagedObjectContext) -> Void) {
         context.perform { [context] in block(context) }
-    }
-
-    public func delete(completion: @escaping DeletionCompletion) {
-        perform { context in
-            do {
-                try ManagedCache.wipe(from: context)
-                try context.save()
-
-                completion(nil)
-            } catch {
-                context.rollback()
-                completion(error)
-            }
-        }
-    }
-
-    public func persist(images: [LocalFeedImage], timestamp: Date, completion: @escaping PersistCompletion) {
-        perform { context in
-            do {
-                try ManagedCache.wipe(from: context)
-
-                let managedCache = ManagedCache(context: context)
-                managedCache.timestamp = timestamp
-                managedCache.feed = ManagedFeedImage.build(with: images, in: context)
-
-                try context.save()
-
-                completion(nil)
-            } catch {
-                context.rollback()
-                completion(error)
-            }
-        }
-
-    }
-
-    public func retrieve(completion: @escaping RetrieveCompletion) {
-        perform { context in
-            do {
-                if let cache = try ManagedCache.find(in: context) {
-                    completion(.found(feed: cache.locals, timestamp: cache.timestamp))
-                } else {
-                    completion(.empty)
-                }
-            } catch {
-                completion(.failure(error))
-            }
-        }
-    }
-
-}
-
-extension CoreDataFeedStore: FeedImageStore {    
-    public func insert(url: URL, with data: Data, completion: @escaping InsertCompletion) {
-        perform { context in
-            do {
-                let model = try ManagedFeedImage.findBy(url: url)
-                model?.data = data
-
-                try context.save()
-                completion(nil)
-            } catch {
-                completion(error)
-            }
-        }
-    }
-
-    public func retrieve(from url: URL, completion: @escaping RetrievalCompletion) {
-        context.perform {
-            do {
-                guard let model = try ManagedFeedImage.findBy(url: url), let imageData = model.data else {
-                    return completion(.empty)
-                }
-
-                completion(.found(imageData))
-            } catch {
-                completion(.failure(error))
-            }
-        }
     }
 }
